+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [1.0.1] - 17.04.2023
+
+### Added
+
+- Added mfplat patch integration
+
+### Fixed
+
+- Fixed launcher pup-up on game launching
+- Fixed game running issue if you have spaces in paths
+
 ## [1.0.0] - 16.04.2023
 
 🚀 Initial release